Add depth_bias field to PolylineMaterial
This let the user draw lines closer or further away from the camera
than the line is really in the 3d world.

A good use case for this is making the lines draw always in front of
geometry, or slightly offsetting the line so they can be shown in front
of geometry it shares with other things.