Fix depth_bias negative values
This commit accounts for homogenous clip space and makes sure the lines
with a -1.0 depth_bias are always shown in front of everything else by
having a depth value exactly equal to the clip's W.

This uses an exponential function for user ergonomics. But honestly, we
could probably do the exponential once on creating the material. and use
linear calculation in the shader.